test(drizzle): move enum filter tests into input-factory.spec.ts
